+++ task/QueryBasicTask.php
@@ -4,7 +4,8 @@ namespace Flex\Banana\Task;
 use Flex\Banana\Classes\TaskFlow;
 use Flex\Banana\Utils\Requested;
 use Flex\Banana\Classes\Db\DbManager;
-
+use Flex\Banana\Classes\Db\DbResultSql;
+use Flex\Banana\Classes\Db\DbResultCouch;
 /**
 $enums = [
     IdEnum::_ID => [],
@@ -22,25 +23,8 @@ class QueryBasicTask
         private array $enums
     ){}
 
-    private function query() : DbManager
+    public function execute(DbResultCouch | DbResultSql $result) : array
     {
-        // 데이터를 쿼리하여 가져옵니다
-        return $this->db->table( $this->task->table )
-            ->select(...array_map(fn($column) => $column->value, $this->task->enums))
-            ->where($this->task->where) // where 조건을 TaskFlow에서 설정된 값으로 사용
-            ->limit($this->task->limit)
-            ->query();
-    }
-
-    private function singleQuery() : void
-    {
-        $result = $this->query();
-        $this->task->data = $result->fetch_assoc();
-    }
-
-    private function multiQuery () : void
-    {
-        $result = $this->query();
         $data = [];
         while ($row = $result->fetch_assoc())
         {
@@ -52,14 +36,6 @@ class QueryBasicTask
             $data[] = $formattedRow;
         }
 
-        $this->task->data = $data;
-    }
-
-    public function execute( string $queryType) : void {
-        if($queryType == 'multi' || $queryType == 'm') {
-            $this->multiQuery();
-        }else{
-            $this->singleQuery();
-        }
+        return $data;
     }
 }